requirements.txt
asgiref==3.7.2
boto3==1.34.55
botocore==1.34.55
certifi==2024.2.2
charset-normalizer==3.3.2
cloudinary==1.39.0
crispy-bootstrap4==2024.1
dj3-cloudinary-storage==0.0.6
Django==5.0.2
django-crispy-forms==2.1
django-storages==1.14.2
gunicorn==21.2.0
idna==3.6
jmespath==1.0.1
mysql-connector==2.2.9
mysqlclient==2.2.4
packaging==23.2
pillow==10.2.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
requests==2.31.0
s3transfer==0.10.0
six==1.16.0
sqlparse==0.4.4
tzdata==2024.1
urllib3==2.0.7
whitenoise==6.6.0
